:tada: 2.7.0.RELEASE,内核全面升级,增加岗位管理,用户导入导出

+++ blade-core-mybatis/src/main/java/org/springblade/core/mp/base/BaseService.java
@@ -34,6 +34,6 @@ public interface BaseService<T> extends IService<T> {
 	 * @param ids id集合(逗号分隔)
 	 * @return boolean
 	 */
-	boolean deleteLogic(@NotEmpty List<Integer> ids);
+	boolean deleteLogic(@NotEmpty List<Long> ids);
 
 }
