+++ blade-core-mybatis/src/main/java/org/springblade/core/mp/base/BaseServiceImpl.java
@@ -65,7 +65,7 @@ public class BaseServiceImpl<M extends BaseMapper<T>, T extends BaseEntity> exte
 	}
 
 	@Override
-	public boolean deleteLogic(@NotEmpty List<Integer> ids) {
+	public boolean deleteLogic(@NotEmpty List<Long> ids) {
 		return super.removeByIds(ids);
 	}
